Made the different module files. The code should still be working properly.
